Extract launch-party components
